requirements.txt
boto3>=1.34
pyyaml>=6.0
pystray>=0.19
Pillow>=10.0
